Deploiement de l authentification

+++ src/main/java/com/locar/dao/AvisRepository.java
@@ -3,5 +3,5 @@ package com.locar.dao;
 import com.locar.entities.Avis;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AvisRepository extends JpaRepository<Avis, Integer>  {
+public interface AvisRepository extends JpaRepository<Avis, Long>  {
 }
