+++ src/main/java/com/locar/dao/CategorieRepository.java
@@ -3,5 +3,5 @@ package com.locar.dao;
 import com.locar.entities.Categorie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategorieRepository extends JpaRepository<Categorie, Integer> {
+public interface CategorieRepository extends JpaRepository<Categorie, Long> {
 }
