+++ src/main/java/com/locar/dao/TarifRepository.java
@@ -3,5 +3,5 @@ package com.locar.dao;
 import com.locar.entities.Tarif;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TarifRepository extends JpaRepository<Tarif, Integer>  {
+public interface TarifRepository extends JpaRepository<Tarif, Long>  {
 }
