+++ src/main/java/com/locar/dao/VehiculeRepository.java
@@ -3,5 +3,5 @@ package com.locar.dao;
 import com.locar.entities.Vehicule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehiculeRepository extends JpaRepository<Vehicule, Integer> {
+public interface VehiculeRepository extends JpaRepository<Vehicule, Long> {
 }
